pip_requirements.txt
Pillow==2.1.0
Scrapy==0.18.2
